studies: import Monk2022 accessions
metadata.tsv in each case is just the list of accessions for the studies.  The
full metadata still needs to be derived from
studies/Monk2022/monk-2022-supplementary-data-1.tsv.  This is enough that we
can start the pipeline, though.

Probably I should go back and change the protocol here so there's a file that's
just a simple list of accessions, since in importing a new study we should
always have that first.