cladecounts: replace allcounts and chidlcounts
Initially I implemented per-taxid counts as two stages, first counting per taxid and then summing per clade.  This is correct for assignments, but not for hits: a read this hit both PMMoV and Tobamovirus should count once for the Tobamovirus clade but the first implementation counted it twice.

Now we run this as a single stage, so we have enough information to count clade hits correctly.