Expand human virus listing
We had been treating human-viruses.tsv as if it was a complete list of taxonomic nodes corresponding to human viruses, but it's actually a list of clades.  This means that if a read had been assigned to, say, 'hepatitis B virus genotype F' we wouldn't have counted it as a human-infecting virus, even though 'Hepatitis B virus' is marked as human-infecting.

This will require regenerating humanviruses/ allmatches/ hvreads/ and the dashboard.