dashboard: start breaking out per-bioproject counts (#40)
As we add more and more data, the dashboard is getting slow to load.  I want to fix this by showing only a subset of the data by default.  To enable loading the data only for specific bioprojects, break comparison_sample_counts and human_virus_sample_counts out by bioproject.

Note that this PR doesn't include updating the dashboard to read these files; that's coming in #41 and #42.