dashboard: separate metadata generation from data processing
Data processing is very slow, but unlikely to fail.  Metadata generation is
much faster, and usually fails a few times when you add a new bioproject
because you haven't got the metadata classifier quite right yet.  Pull these
into two separate scripts, and move metadata generation as early as possible.

Note that this does mean samples will be included in the sample list even if we
they haven't finished processing, but I think this is fine.